move newrelic integration from subpackage to main package
