+++ efopen/ef_generate.py
@@ -554,7 +554,8 @@ def create_newrelic_alerts():
    Note: Import is inside this function rather than top of the file so that we do not import when running unit tests.
          (would otherwise require an ef_site_config.yml in the directory where tests are being run)
   """
-  from efopen.newrelic.executor import NewRelicAlerts
+  from newrelic_executor import NewRelicAlerts
+  print("> Creating NewRelic Alerts")
   NewRelicAlerts(CONTEXT, CLIENTS).run()
  
 def main():
